PID Goes really well.
Calibrated with Integral error limits.
Derivative term seems to do nothing. Use float or a scaled long? The number updates very rarely, so it's mostly 0 or noise.
I want to do a test, where you push the throttle and it has a lot of damping in one direction, but 0 in the other.