Move all global variables to global.c and global.h
The variables are defined (reserved memory for them) in global.c, and elsewhere in the files they shall be declared to be used locally with the keyword "extern".
To do that easily, #include the file "global.h". In that way, when we change the type or name of a global variable, we won't have to edit every single file where it is used, but just in global.c and global.h. It's less prone to errors.